requirements.txt
numpy
torch
torchvision
torchaudio
faiss-cpu
matplotlib
tqdm
ipykernal
pyyaml